+++ pkg/grpc/default_api/api_state_test.go
@@ -14,11 +14,119 @@
 package default_api
 
 import (
+	"context"
+	"fmt"
 	"github.com/dapr/components-contrib/state"
+	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
+	mock_state "mosn.io/layotto/pkg/mock/components/state"
 	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 	"testing"
 )
+
+func TestSaveState(t *testing.T) {
+	t.Run("normal", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().BulkSet(gomock.Any()).DoAndReturn(func(reqs []state.SetRequest) error {
+			assert.Equal(t, 1, len(reqs))
+			assert.Equal(t, "abc", reqs[0].Key)
+			assert.Equal(t, []byte("mock data"), reqs[0].Value)
+			return nil
+		})
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.SaveStateRequest{
+			StoreName: "mock",
+			States: []*runtimev1pb.StateItem{
+				{
+					Key:   "abc",
+					Value: []byte("mock data"),
+				},
+			},
+		}
+		_, err := api.SaveState(context.Background(), req)
+		assert.Nil(t, err)
+	})
+	t.Run("with options last-write and eventual", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().BulkSet(gomock.Any()).DoAndReturn(func(reqs []state.SetRequest) error {
+			assert.Equal(t, 1, len(reqs))
+			assert.Equal(t, "abc", reqs[0].Key)
+			assert.Equal(t, []byte("mock data"), reqs[0].Value)
+			assert.Equal(t, "last-write", reqs[0].Options.Concurrency)
+			assert.Equal(t, "eventual", reqs[0].Options.Consistency)
+			return nil
+		})
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.SaveStateRequest{
+			StoreName: "mock",
+			States: []*runtimev1pb.StateItem{
+				{
+					Key:   "abc",
+					Value: []byte("mock data"),
+					Options: &runtimev1pb.StateOptions{
+						Concurrency: runtimev1pb.StateOptions_CONCURRENCY_LAST_WRITE,
+						Consistency: runtimev1pb.StateOptions_CONSISTENCY_EVENTUAL,
+					},
+				},
+			},
+		}
+		_, err := api.SaveState(context.Background(), req)
+		assert.Nil(t, err)
+	})
+	t.Run("with options first-write and strong", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().BulkSet(gomock.Any()).DoAndReturn(func(reqs []state.SetRequest) error {
+			assert.Equal(t, 1, len(reqs))
+			assert.Equal(t, "abc", reqs[0].Key)
+			assert.Equal(t, []byte("mock data"), reqs[0].Value)
+			assert.Equal(t, "first-write", reqs[0].Options.Concurrency)
+			assert.Equal(t, "strong", reqs[0].Options.Consistency)
+			return nil
+		})
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.SaveStateRequest{
+			StoreName: "mock",
+			States: []*runtimev1pb.StateItem{
+				{
+					Key:   "abc",
+					Value: []byte("mock data"),
+					Options: &runtimev1pb.StateOptions{
+						Concurrency: runtimev1pb.StateOptions_CONCURRENCY_FIRST_WRITE,
+						Consistency: runtimev1pb.StateOptions_CONSISTENCY_STRONG,
+					},
+				},
+			},
+		}
+		_, err := api.SaveState(context.Background(), req)
+		assert.Nil(t, err)
+	})
+
+	t.Run("save error", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		mockStore := mock_state.NewMockStore(ctrl)
+		mockStore.EXPECT().Features().Return(nil)
+		mockStore.EXPECT().BulkSet(gomock.Any()).Return(fmt.Errorf("net error"))
+		api := NewAPI("", nil, nil, nil, nil, map[string]state.Store{"mock": mockStore}, nil, nil, nil, nil)
+		req := &runtimev1pb.SaveStateRequest{
+			StoreName: "mock",
+			States: []*runtimev1pb.StateItem{
+				{
+					Key:   "abc",
+					Value: []byte("mock data"),
+				},
+			},
+		}
+		_, err := api.SaveState(context.Background(), req)
+		assert.NotNil(t, err)
+		assert.Equal(t, "rpc error: code = Internal desc = failed saving state in state store mock: net error", err.Error())
+	})
+}
 
 func TestGetResponse2GetStateResponse(t *testing.T) {
 	resp := GetResponse2GetStateResponse(&state.GetResponse{
